+++ pkg/config/compiler_validate_strict_observability.go
@@ -1,0 +1,531 @@
+package config
+
+import (
+	"fmt"
+	"sort"
+	"strings"
+)
+
+// validateLogEventModeFormatStrict rejects a top-level `security log format`
+// that the EVENT-mode writer cannot honor (#3349 follow-up). The top-level
+// format leaf is schema-validated to one of {binary, sd-syslog, structured,
+// syslog} regardless of mode, and BOTH runtimes now honor every one of those
+// values (#3409 closed the event-mode gap):
+//
+//   - stream mode (remote syslog): honors binary (formatBinaryRecord),
+//     structured (formatStructuredMsg), sd-syslog (RFC 5424 envelope in
+//     SyslogClient.Send), and the standard RFC 3164 default.
+//   - event mode (local file, pkg/logging LocalLogWriter via
+//     ringbuf.go ProcessRawEvent local-writer fanout): honors binary
+//     (formatBinaryRecord), structured (Junos RT_FLOW body), sd-syslog
+//     (RFC 5424 envelope in LocalLogWriter.Send), and the standard default.
+//
+// Before #3409 the event-mode fanout branched ONLY on `binary` and wrote
+// standard text for every other value, so `structured` and `sd-syslog`
+// SILENTLY fell back to standard text — the exact silent-config-fallback #3349
+// exists to eliminate. That gap is closed: each format now produces distinct
+// output, so every schema value is accepted in event mode too.
+//
+// This stays as a CROSS-FIELD validator (format validity in principle depends
+// on the sibling mode, which the declarative per-leaf SchemaValidate walker
+// cannot express) so the contract is documented in one place and a future
+// event-only format remains easy to gate. Today it accepts the full schema
+// enum for either mode.
+//
+// On the tolerant load / peer-sync paths the call site still downgrades any
+// rejection to a warning (opts.lenientLogEventModeFormat); with the full enum
+// accepted this path is now inert for the four known formats and only fires if
+// a future value is added to the schema but not yet honored here.
+//
+// The event-honorable set MUST stay in sync with the LocalLogWriter fanout in
+// pkg/logging/ringbuf.go: a value allowed here but unhonored there would
+// reintroduce the silent fallback.
+func validateLogEventModeFormatStrict(cfg *Config) error {
+	if cfg == nil || cfg.Security.Log.Mode != "event" {
+		return nil
+	}
+	switch cfg.Security.Log.Format {
+	case "", "binary", "syslog", "structured", "sd-syslog":
+		// All four schema formats are honored by the event-mode LocalLogWriter
+		// fanout (#3409): "" / "syslog" => standard RFC 3164 text; "binary" =>
+		// binary records; "structured" => Junos RT_FLOW body; "sd-syslog" =>
+		// RFC 5424 envelope. Nothing silently falls back.
+		return nil
+	default:
+		// Defensive: the schema leaf already constrains the value to the set
+		// above. An unknown value reaching here would be emitted as standard
+		// text by the writer, so reject it rather than silently no-op — the
+		// #3349 contract.
+		return fmt.Errorf("security log format %q is not honored in event mode "+
+			"(the local-file writer emits binary, structured, sd-syslog, or "+
+			"standard text); use one of those formats", cfg.Security.Log.Format)
+	}
+}
+
+// validateLogProfileStreamReferencesStrict hard-rejects a
+// `security log profile <name>` whose `stream-name` reference does not
+// resolve to a configured `security log stream` (#2008 H7). xpf routes
+// log events per stream (a Junos superset — every matching stream
+// receives the event), so a profile's `stream-name` designates the
+// stream that carries its events. A profile naming a stream that is not
+// configured routes to nowhere: the operator authored a log profile
+// whose target silently never fires. Before H7 the whole profile stanza
+// was dropped before compile, so the typo was invisible; now the
+// reference is validated.
+//
+// A profile with no `stream-name` is accepted: Junos permits a profile
+// that relies on the global routing inheritance, and there is nothing to
+// dangle. Only a non-empty `stream-name` that misses the stream map is
+// rejected.
+//
+// Note: compileLog only records a stream in Log.Streams when it has a
+// host (a host-less stream is not a real destination and is dropped by
+// the stream loop), so a profile referencing a host-less stream is
+// treated as a dangling reference — consistent with the stream's own
+// "must have a host to exist" semantics.
+//
+// On the tolerant load / peer-sync paths the call site downgrades this
+// to a warning (opts.lenientLogProfileStreamRef) so an already-persisted
+// config (older binaries dropped the stanza entirely) or a peer-synced
+// config still boots. Mirrors validateIPsecPolicyProposalReferencesStrict.
+func validateLogProfileStreamReferencesStrict(cfg *Config) error {
+	if cfg == nil {
+		return nil
+	}
+	profiles := cfg.Security.Log.Profiles
+	if len(profiles) == 0 {
+		return nil
+	}
+	streams := cfg.Security.Log.Streams
+	// Profiles is a map (unordered); sort keys so the first-error
+	// commit-check message is deterministic across runs.
+	names := make([]string, 0, len(profiles))
+	for name := range profiles {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+	for _, name := range names {
+		p := profiles[name]
+		if p == nil || p.StreamName == "" {
+			continue
+		}
+		if _, ok := streams[p.StreamName]; ok {
+			continue
+		}
+		return fmt.Errorf("security log profile %q references undefined "+
+			"log stream %q (the profile would route to nowhere — define "+
+			"the stream or fix the stream-name)", p.Name, p.StreamName)
+	}
+	return nil
+}
+
+// validateDynamicAddressFeedServerEndpointStrict hard-rejects a
+// `security dynamic-address feed-server <name>` that configures no endpoint —
+// neither `url` nor `hostname` (#3300 residual). feeds.Manager.Apply derives
+// each server's base URL via resolveBaseURL (feeds.go): explicit `url`, else
+// `https://<hostname>`, else the empty string — and an empty base URL makes
+// Apply SKIP the whole server (it registers NONE of its feeds, including any
+// nested feed-name entries). The endpoint-less server still compiles into
+// SecurityConfig.DynamicAddress.FeedServers, so its feed names are
+// syntactically "declared", but at runtime no feed exists: an address-name
+// bound to one resolves to an empty (match-nothing) address book and a
+// feed-backed deny policy silently denies nothing — the same #3300 fail-open
+// symptom one layer up, at the feed-server root rather than the binding.
+//
+// This gate replicates resolveBaseURL's emptiness condition directly on the
+// FeedServer config struct (feedServerBaseURLEmpty) rather than importing
+// pkg/feeds (pkg/config must not depend on pkg/feeds). resolveBaseURL prefers
+// `url` and returns strings.TrimRight(url, "/") BEFORE it ever falls back to
+// `hostname`, so a slash-only `url` (e.g. `/`, `//`) trims to "" and the
+// server is skipped even when a hostname is also set — feedServerBaseURLEmpty
+// mirrors that branch order exactly. Keep in sync with resolveBaseURL.
+//
+// On the tolerant load / peer-sync paths the call site downgrades this to a
+// warning (opts.lenientDynamicAddressFeedRef, shared with the feed-name
+// cross-reference gate) so an already-persisted or peer-synced config carrying
+// an endpoint-less server still boots — the runtime already drops the server
+// (registers no feed), so any bound address-name is fail-closed match-none
+// rather than bricking the load (#1960 / #3261 class). Commit / commit-check
+// stay strict. Mirrors validateLogProfileStreamReferencesStrict.
+func validateDynamicAddressFeedServerEndpointStrict(cfg *Config) error {
+	if cfg == nil {
+		return nil
+	}
+	servers := cfg.Security.DynamicAddress.FeedServers
+	if len(servers) == 0 {
+		return nil
+	}
+	// FeedServers is a map (unordered); sort keys so the first-error
+	// commit-check message is deterministic across runs.
+	names := make([]string, 0, len(servers))
+	for name := range servers {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+	for _, name := range names {
+		fs := servers[name]
+		if fs == nil {
+			continue
+		}
+		if feedServerBaseURLEmpty(fs) {
+			display := fs.Name
+			if display == "" {
+				display = name
+			}
+			return fmt.Errorf("security dynamic-address feed-server %q resolves "+
+				"to an empty endpoint (no url or hostname, or a slash-only url) "+
+				"so it registers no feeds — any address-name bound to it "+
+				"silently matches nothing; set a valid url or hostname",
+				display)
+		}
+	}
+	return nil
+}
+
+// feedServerBaseURLEmpty reports whether feeds.resolveBaseURL would return ""
+// for this feed-server — i.e. feeds.Manager.Apply would SKIP it and register
+// none of its feeds. It mirrors resolveBaseURL (pkg/feeds/feeds.go)
+// BRANCH-FOR-BRANCH:
+//
+//	if URL != "":            empty iff strings.TrimRight(URL, "/") == ""
+//	else if Hostname != "":  never empty ("https://" + ... is always non-empty)
+//	else:                    empty
+//
+// The URL branch wins outright, so a slash-only `url` (e.g. `/`, `//`) trims to
+// "" and the server is skipped EVEN IF a hostname is also configured — the
+// fallback is never reached. resolveBaseURL performs no whitespace trimming
+// (only strings.TrimRight on "/"), so this does not either. pkg/config cannot
+// import pkg/feeds (import cycle); keep this in sync with resolveBaseURL.
+func feedServerBaseURLEmpty(fs *FeedServer) bool {
+	if fs.URL != "" {
+		return strings.TrimRight(fs.URL, "/") == ""
+	}
+	if fs.Hostname != "" {
+		return false
+	}
+	return true
+}
+
+// validateDynamicAddressFeedReferencesStrict hard-rejects a
+// `security dynamic-address address-name <addr> profile feed-name <feed>`
+// binding whose `<feed>` resolves to no declared feed (#3300). The
+// address-name→feed binding is recorded verbatim into
+// AddressBinding.FeedNames with no cross-reference against the configured
+// feed-servers (compileDynamicAddress in compiler_services.go), and at
+// runtime an unknown feed name contributes nothing: feeds.Manager.
+// SnapshotForBindings skips a feed it never registered and still publishes
+// a non-nil EMPTY prefix set for the binding (feeds.go SnapshotForBindings),
+// so the AF_XDP address book gets a book ID that matches nothing
+// (fail-closed). The runtime fail-closed posture is correct for "feed
+// declared but not yet fetched", but a TYPO in the feed-name is
+// indistinguishable from that and arms a silent match-none book: a
+// feed-backed deny policy referencing the dynamic-address denies nothing,
+// with no commit error. Junos rejects an address-name whose profile
+// feed-name does not resolve to a declared feed at commit; this gate
+// restores that behavior.
+//
+// The valid feed-name set mirrors the keys feeds.Manager registers (feeds.go
+// Start): a feed-server with per-feed entries contributes each FeedEntry.Name;
+// a single-feed server contributes its FeedName, or the server name itself
+// when no explicit feed-name is set. This parity is EXACT because
+// validateDynamicAddressFeedServerEndpointStrict (run just before this gate)
+// rejects any feed-server with no url/hostname — the only servers
+// feeds.Manager.Apply silently SKIPS — so every server in the declared set is
+// one Apply would actually register. The schema accepts `profile feed-name` as
+// a free-form value leaf (schema_security.go), so the undefined-token gate
+// (#2008/#2009) does NOT cover a typo here — only this cross-reference does.
+//
+// On the tolerant load / peer-sync paths the call site downgrades this to a
+// warning (opts.lenientDynamicAddressFeedRef) so an already-persisted config
+// (older binaries never validated the reference) or a peer-synced config
+// still boots — the runtime is already fail-closed (match-none) for an
+// unknown feed, so a leniently-loaded typo denies nothing rather than
+// bricking the load (#1960 / #3261 class). Commit / commit-check stay strict
+// so the operator's next edit fails loudly. Mirrors
+// validateLogProfileStreamReferencesStrict.
+func validateDynamicAddressFeedReferencesStrict(cfg *Config) error {
+	if cfg == nil {
+		return nil
+	}
+	da := cfg.Security.DynamicAddress
+	if len(da.AddressBindings) == 0 {
+		return nil
+	}
+
+	// Build the set of declared feed names exactly as feeds.Manager keys
+	// them (pkg/feeds Start): FeedEntries name each feed; a single-feed
+	// server keys on FeedName, falling back to the server name.
+	declared := make(map[string]bool)
+	for _, fs := range da.FeedServers {
+		if fs == nil {
+			continue
+		}
+		if len(fs.FeedEntries) > 0 {
+			for _, fe := range fs.FeedEntries {
+				declared[fe.Name] = true
+			}
+			continue
+		}
+		key := fs.FeedName
+		if key == "" {
+			key = fs.Name
+		}
+		if key != "" {
+			declared[key] = true
+		}
+	}
+
+	// AddressBindings is a map (unordered); sort keys so the first-error
+	// commit-check message is deterministic across runs.
+	names := make([]string, 0, len(da.AddressBindings))
+	for name := range da.AddressBindings {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+	for _, name := range names {
+		ab := da.AddressBindings[name]
+		if ab == nil {
+			continue
+		}
+		for _, fn := range ab.FeedNames {
+			if fn == "" || declared[fn] {
+				continue
+			}
+			return fmt.Errorf("security dynamic-address address-name %q "+
+				"profile references undefined feed-name %q (the binding would "+
+				"resolve to an empty address set — a feed-backed policy would "+
+				"silently match nothing; declare the feed under a "+
+				"dynamic-address feed-server or fix the feed-name)", ab.Name, fn)
+		}
+	}
+	return nil
+}
+
+// validateFlowServerTemplateReferencesStrict hard-rejects a per-flow-server
+// NetFlow v9 / IPFIX template reference (`version9 { template <name> }`,
+// `version9-template <name>`, `version-ipfix { template <name> }`, or
+// `version-ipfix-template <name>`) that names no template defined under the
+// matching `services flow-monitoring` version stanza (#2461).
+//
+// Without this gate the live exporter (pkg/flowexport) ignored the per-server
+// reference entirely and built one export config from the FIRST Go-map-
+// iteration template, broadcasting it to every collector of that version. A
+// collector that asked for a specific template silently received whichever
+// template the map happened to yield first, and that choice flipped across
+// process restarts (map order is not an operator contract). A reference to a
+// template that does not exist at all is the clearest form of the same defect:
+// the operator's intent (timeouts / export-extensions) is simply dropped.
+//
+// On the tolerant load / peer-sync paths the call site downgrades this to a
+// warning (opts.lenientFlowServerTemplateRef) so an already-persisted or
+// peer-synced config carrying the typo still boots (#1960 fail-closed-on-load
+// class). The resolver (ResolveV9TemplateGroups / ResolveIPFIXTemplateGroups)
+// drops a group whose referenced template is undefined, so a leniently-loaded
+// bad config exports nothing for that collector rather than the wrong
+// template. Commit / commit-check stay strict so the operator's next edit
+// fails loudly. Mirrors validateLogProfileStreamReferencesStrict.
+func validateFlowServerTemplateReferencesStrict(cfg *Config) error {
+	if cfg == nil {
+		return nil
+	}
+	fm := cfg.Services.FlowMonitoring
+	if cfg.ForwardingOptions.Sampling == nil {
+		return nil
+	}
+
+	v9Defined := map[string]bool{}
+	ipfixDefined := map[string]bool{}
+	if fm != nil {
+		if fm.Version9 != nil {
+			for name := range fm.Version9.Templates {
+				v9Defined[name] = true
+			}
+		}
+		if fm.VersionIPFIX != nil {
+			for name := range fm.VersionIPFIX.Templates {
+				ipfixDefined[name] = true
+			}
+		}
+	}
+
+	// Walk the sampling instances in a deterministic key order so the
+	// first-error commit-check message is stable across runs (the instance
+	// map is unordered).
+	instNames := make([]string, 0, len(cfg.ForwardingOptions.Sampling.Instances))
+	for name := range cfg.ForwardingOptions.Sampling.Instances {
+		instNames = append(instNames, name)
+	}
+	sort.Strings(instNames)
+
+	for _, instName := range instNames {
+		inst := cfg.ForwardingOptions.Sampling.Instances[instName]
+		if inst == nil {
+			continue
+		}
+		for _, fam := range []*SamplingFamily{inst.FamilyInet, inst.FamilyInet6} {
+			if fam == nil {
+				continue
+			}
+			for _, fs := range fam.FlowServers {
+				if fs == nil {
+					continue
+				}
+				if fs.Version9Template != "" && !v9Defined[fs.Version9Template] {
+					return fmt.Errorf("forwarding-options sampling instance %q "+
+						"flow-server %s references undefined version9 template %q "+
+						"(define it under services flow-monitoring version9, or fix "+
+						"the template name — the collector would otherwise receive "+
+						"an arbitrary template)", instName, fs.Address, fs.Version9Template)
+				}
+				if fs.VersionIPFIXTemplate != "" && !ipfixDefined[fs.VersionIPFIXTemplate] {
+					return fmt.Errorf("forwarding-options sampling instance %q "+
+						"flow-server %s references undefined version-ipfix template "+
+						"%q (define it under services flow-monitoring version-ipfix, "+
+						"or fix the template name — the collector would otherwise "+
+						"receive an arbitrary template)", instName, fs.Address, fs.VersionIPFIXTemplate)
+				}
+			}
+		}
+	}
+	return nil
+}
+
+// flowServerExportVersion mirrors flowexport.resolveFlowServerVersion: it
+// returns the export protocol a single flow-server binds to, given the
+// per-server selector and which global `services flow-monitoring` version
+// stanzas are configured. It is duplicated here (not imported) because
+// pkg/config must not depend on pkg/flowexport. Returns "" when the server
+// resolves to no configured version. Keep in sync with
+// pkg/flowexport.resolveFlowServerVersion.
+func flowServerExportVersion(fs *FlowServer, hasV9, hasIPFIX bool) string {
+	switch fs.Version {
+	case FlowServerVersion9:
+		if hasV9 {
+			return FlowServerVersion9
+		}
+		return ""
+	case FlowServerVersionIPFIX:
+		if hasIPFIX {
+			return FlowServerVersionIPFIX
+		}
+		return ""
+	}
+	switch {
+	case hasIPFIX:
+		return FlowServerVersionIPFIX
+	case hasV9:
+		return FlowServerVersion9
+	default:
+		return ""
+	}
+}
+
+// validateSamplingInstanceConflictsStrict hard-rejects an unsupported
+// multi-sampling-instance configuration (#2462).
+//
+// The defect: multiple `forwarding-options sampling instance` blocks were
+// silently flattened into one global export policy — one rate (the first
+// nonzero InputRate in Go map order), one merged collector set, one zone
+// eligibility map. Flows from instance A could export to instance B's
+// collectors and the effective rate depended on map-iteration order.
+//
+// The fix makes each instance a first-class export policy: its own rate, its
+// own 1-in-N counter, its own collectors. A flow is attributed to an instance
+// by ADDRESS FAMILY (the only per-flow selector available — the interface
+// `family inet { sampling { input; } }` stanza is a plain boolean; there is
+// NO per-interface sampling-instance selector in the config model, so two
+// instances serving the SAME family for the SAME export version are genuinely
+// ambiguous: the runtime cannot tell which instance a given IPv4 (or IPv6)
+// flow belongs to). Rather than guess (the flatten-and-hope behavior this
+// issue reports), that combination is rejected.
+//
+// Supported: a single instance (any rate / families / collectors — the common
+// case, unchanged); multiple instances disambiguated by family (e.g. instance
+// A serves inet, instance B serves inet6) and/or by export version (an inet
+// instance bound to version9 vs an inet instance bound to version-ipfix —
+// distinct datagram streams, the flow is duplicated to both intentionally,
+// same as today's single-instance dual-version behavior).
+//
+// Rejected: two or more instances each configuring a flow-server that
+// resolves to the SAME (export-version, address-family) pair.
+//
+// Strict on commit / commit-check (hard reject so the operator sees it);
+// lenient on load / peer-sync (the call site downgrades to a warning via
+// opts.lenientSamplingInstanceConflicts so an already-persisted or
+// peer-synced config still boots — #1960; the resolver still emits both
+// instances' independent ExportConfigs, so a leniently-loaded conflicting
+// config duplicates eligible flows to both instances rather than bricking).
+func validateSamplingInstanceConflictsStrict(cfg *Config) error {
+	if cfg == nil || cfg.ForwardingOptions.Sampling == nil {
+		return nil
+	}
+	insts := cfg.ForwardingOptions.Sampling.Instances
+	if len(insts) < 2 {
+		return nil // single instance is always unambiguous
+	}
+	fm := cfg.Services.FlowMonitoring
+	hasV9 := fm != nil && fm.Version9 != nil
+	hasIPFIX := fm != nil && fm.VersionIPFIX != nil
+
+	// Deterministic instance order so the first-conflict message is stable.
+	names := make([]string, 0, len(insts))
+	for name := range insts {
+		names = append(names, name)
+	}
+	sort.Strings(names)
+
+	// claim key "<version>\x00<family>" -> first instance that claimed it.
+	claimed := map[string]string{}
+	for _, name := range names {
+		inst := insts[name]
+		if inst == nil {
+			continue
+		}
+		fams := []struct {
+			fam    *SamplingFamily
+			family string
+		}{
+			{inst.FamilyInet, "inet"},
+			{inst.FamilyInet6, "inet6"},
+		}
+		// Collect THIS instance's distinct (version, family) claims, so the
+		// same instance binding two collectors to the same (version, family)
+		// does not self-conflict.
+		selfClaims := map[string]bool{}
+		for _, fe := range fams {
+			if fe.fam == nil {
+				continue
+			}
+			for _, fs := range fe.fam.FlowServers {
+				if fs == nil {
+					continue
+				}
+				ver := flowServerExportVersion(fs, hasV9, hasIPFIX)
+				if ver == "" {
+					continue // resolves to no configured version; exports nothing
+				}
+				selfClaims[ver+"\x00"+fe.family] = true
+			}
+		}
+		// Deterministic order over this instance's claims.
+		keys := make([]string, 0, len(selfClaims))
+		for k := range selfClaims {
+			keys = append(keys, k)
+		}
+		sort.Strings(keys)
+		for _, key := range keys {
+			if owner, ok := claimed[key]; ok {
+				parts := strings.SplitN(key, "\x00", 2)
+				return fmt.Errorf("forwarding-options sampling instances %q and %q "+
+					"both export %s family %s flows: the runtime cannot attribute a "+
+					"flow to one instance (there is no per-interface sampling-instance "+
+					"selector — eligibility is per address family), so the two would "+
+					"silently merge. Use a single instance for this version/family, or "+
+					"separate the instances by family or export version",
+					owner, name, parts[0], parts[1])
+			}
+			claimed[key] = name
+		}
+	}
+	return nil
+}
